+++ koodisto-api/src/main/java/fi/vm/sade/koodisto/service/types/common/KoodiMetadataType.java
@@ -3,15 +3,10 @@ package fi.vm.sade.koodisto.service.types.common;
 
 import java.io.Serializable;
 
-public class KoodiMetadataType implements Serializable
+public class KoodiMetadataType extends MetadataType implements Serializable
 {
 
     private static final long serialVersionUID = 1L;
-
-    protected KieliType kieli;
-
-    protected String nimi;
-    protected String kuvaus;
     protected String lyhytNimi;
     protected String kayttoohje;
     protected String kasite;
@@ -19,30 +14,6 @@ public class KoodiMetadataType implements Serializable
     protected String eiSisallaMerkitysta;
     protected String huomioitavaKoodi;
     protected String sisaltaaKoodiston;
-
-    public KieliType getKieli() {
-        return kieli;
-    }
-
-    public void setKieli(KieliType value) {
-        this.kieli = value;
-    }
-
-    public String getNimi() {
-        return nimi;
-    }
-
-    public void setNimi(String value) {
-        this.nimi = value;
-    }
-
-    public String getKuvaus() {
-        return kuvaus;
-    }
-
-    public void setKuvaus(String value) {
-        this.kuvaus = value;
-    }
 
     public String getLyhytNimi() {
         return lyhytNimi;
