+++ koodisto-api/src/main/java/fi/vm/sade/koodisto/service/types/common/KoodistoMetadataType.java
@@ -3,13 +3,10 @@ package fi.vm.sade.koodisto.service.types.common;
 
 import java.io.Serializable;
 
-public class KoodistoMetadataType implements Serializable
+public class KoodistoMetadataType extends MetadataType implements Serializable
 {
 
     private static final long serialVersionUID = 1L;
-    protected KieliType kieli;
-    protected String nimi;
-    protected String kuvaus;
     protected String kayttoohje;
     protected String kasite;
     protected String kohdealue;
@@ -19,30 +16,6 @@ public class KoodistoMetadataType implements Serializable
     protected String tarkentaaKoodistoa;
     protected String huomioitavaKoodisto;
     protected String koodistonLahde;
-
-    public KieliType getKieli() {
-        return kieli;
-    }
-
-    public void setKieli(KieliType value) {
-        this.kieli = value;
-    }
-
-    public String getNimi() {
-        return nimi;
-    }
-
-    public void setNimi(String value) {
-        this.nimi = value;
-    }
-
-    public String getKuvaus() {
-        return kuvaus;
-    }
-
-    public void setKuvaus(String value) {
-        this.kuvaus = value;
-    }
 
     public String getKayttoohje() {
         return kayttoohje;
